Moved controllers, deleted issuer controller for now, improved patterns
